[fix] force hps matrix to be the correct size
Using squeeze to remove the first singleton dimension fails when the
first two dimensions are singleton; in this case, squeeze returns a
column vector instead of a row vector, which messes with other functions
down the line. Using reshape instead, we can just remove the first
singleton dimension, which will work in all cases.