split training data for sparse coding
The dictionary needs to be trained on non-insect images, and those
images should not be used in any other training or testing, so as to not
bias the results. For now, we are using 10% of the non-insect images to
train the dictionary.

The original training data split has been renamed so the remaining 90%
of the training data can just be called 'trainingData', since that is
what all normal training will happen on.