[fix] force fundamental to be the right size
Using find() on a matrix doesn't return values for rows that don't have
any non-zero elements in them. This caused the fundamental vector
to be the wrong size. Using a for loop fixes that issue. When a maxima
isn't found, the fundamental location is set to 1, indicating that no
peak was found; setting it to 1 results in the frequency bin being 0,
which should force all the harmonic frequencies to be 0 (hopefully).