format data and labels for classifiers
The classifiers need the data and labels in a different format than
how we store the data. In the case of the Stats and ML Toolbox, which
this commit deals with, the classifiers need everything as a single
matrix or table, not cell arrays; the Deep Learning Toolbox has
different formatting requirements. StatsToolboxClassifier was created to
handle formatting the data and labels for the Stats Toolbox classifiers;
a different subclass can be created for the Deep Learning Toolbox.