add name property to classifiers
When saving files specific to each classifier, we need to add the
classifier name to the filename. Using class() doesn't work because we
only have access to the function handle of the classifier when saving
the grid search results. Additionally, SVM can support multiple kernel
functions, each of which would be considered a separate classifier for
model selection purposes; thus SVM's classifier name will not just be
the class name.