add bandpass filter before frequency feature extraction
Bandpass filtering around the wingbeat frequencies should help
distinguish between insects and non-insects, as well as get rid of the
mostly-DC isnect-body signal contributions, which tend to mess up our
frequency estimation. Now that the filter eliminates most of the DC
components, we don't want to normalize the ESD by the DC component.