add support for optimizable params being a struct
The optimizable parameters were saved as a struct, not a table,
after hyperparameter tuning. The saved hyperparameters still need
to be formatted for the classifier constructors for the final
training, though. Thus we add support for formatting optimizable
parameters that were passed as a struct. Alternatively, the parameters
could (and should) be saved in the format the classifier wants after
hyperparameter tuning, instead.