seed the random number generator after data sampling
The data sampling function seeds the random number generator when doing
undersampling and oversampling. We want to reseed it right before Bayesian
optimization to ensure that the Bayesian optimization is reproducible.
By seeding the random number generator right before the Bayesian
optimization is performed, the initial seed points should always be the
same.